Reorganise traits
