+++ src/eeprom/file_reader.rs
@@ -1,0 +1,3 @@
+//! An EEPROM reader backed by an EEPROM image file instead of a real device.
+//!
+//! Useful for debugging and unit testing.
